fixed interfaces,moved about in component App
