--- src/About.css
@@ -1,6 +1,0 @@
-.about {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 100vh;
-}
