+++ src/App.css
@@ -1,4 +1,7 @@
-body {
+.about {
   display: flex;
   flex-direction: column;
+  justify-content: space-between;
+  min-height: 100vh;
 }
+
