+++ src/App.tsx
@@ -1,10 +1,16 @@
-import About from '@/About.tsx';
+import AboutComponent from '@/components/about/About.component.tsx';
+import FooterComponent from '@/components/footer/Footer.component.tsx';
+import HeaderComponent from '@/components/header/Header.component.tsx';
+
+import './App.css';
 
 function App() {
     return (
-        <>
-            <About />
-        </>
+        <div className="about">
+            <HeaderComponent />
+            <AboutComponent />
+            <FooterComponent />
+        </div>
     );
 }
