+++ src/interface/button.ts
@@ -1,0 +1,9 @@
+interface ButtonInterface {
+    name?: string;
+    backgroundColor?: string;
+    textColor?: string;
+    icon?: string;
+    className?: string;
+}
+
+export default ButtonInterface;
